Added invoice functions

+++ Expedia challenge/InvoiveHandler/src/main/java/com/practise/invoice/Calculator.java
@@ -1,7 +1,7 @@
 package com.practise.invoice;
 
 public interface Calculator {
-    public float getItemFinalPriceIncludingTax(Item item);
-    public float getFinalPurchasedPrice(Item item);
-    public float applyRegularTax(Item item);
+    public double getItemFinalPriceIncludingTax(Item item);
+    public double getFinalPurchasedPrice(Item item);
+    public double applyRegularTax(Item item);
 }
